refactor(middleware-ace): extract teardown into own module + reduce v6 migration
Move runTeardown + helpers from pipeline-helpers.ts to a new teardown.ts so
pipeline-helpers stays under the 400-line soft limit. Refactor
migrateWatermarksToV6 by extracting collectV6Watermarks helper to drop below
the 50-line function cap. Replace a non-null assertion in runStructuredPipeline
with sentinel-based min/max computation (no behavior change — empty entries
already early-returns above).

Net effect: complexity ratchet at 779 (cap 779) — passing.
